+++ packages/plugins/learn-card-network/src/plugin.ts
@@ -696,6 +696,16 @@ export const getLearnCardNetworkPlugin = async (
 
                 return client.profile.signingAuthority.query({ endpoint, name });
             },
+            setPrimaryRegisteredSigningAuthority: async (_learnCard, endpoint, name) => {
+                if (!userData) throw new Error('Please make an account first!');
+
+                return client.profile.setPrimarySigningAuthority.mutate({ endpoint, name });
+            }, 
+            getPrimaryRegisteredSigningAuthority: async (_learnCard) => {
+                if (!userData) throw new Error('Please make an account first!');
+
+                return client.profile.primarySigningAuthority.query();
+            }, 
 
             generateClaimLink: async (_learnCard, boostUri, claimLinkSA, options, challenge) => {
                 if (!userData) throw new Error('Please make an account first!');
@@ -960,6 +970,40 @@ export const getLearnCardNetworkPlugin = async (
                 if (!apiToken) throw new Error('Failed to get API Token for auth grant');
 
                 return apiToken;
+            },
+
+
+            sendCredentialViaInbox: async (_learnCard, issueInboxCredential) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.inbox.issue.mutate(issueInboxCredential);
+            },
+            getMySentInboxCredentials: async (_learnCard, options) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.inbox.getMyIssuedCredentials.query(options);
+            },
+            getInboxCredential: async (_learnCard, id) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.inbox.getInboxCredential.query({ credentialId: id });
+            },
+            addContactMethod: async (_learnCard, contactMethod) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.contactMethods.addContactMethod.mutate(contactMethod);
+            },
+            getMyContactMethods: async (_learnCard) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.contactMethods.getMyContactMethods.query();
+            },
+            setPrimaryContactMethod: async (_learnCard, contactMethodId) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.contactMethods.setPrimaryContactMethod.mutate({ contactMethodId });
+            },
+            verifyContactMethod: async (_learnCard, token) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.contactMethods.verifyContactMethod.mutate({ token });
+            },
+            removeContactMethod: async (_learnCard, id) => {
+                if (!userData) throw new Error('Please make an account first!');
+                return client.contactMethods.removeContactMethod.mutate({ id });
             },
 
             resolveFromLCN: async (_learnCard, uri) => {