+++ packages/plugins/claimable-boosts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @learncard/claimable-boosts-plugin
 
+## 1.0.41
+
+### Patch Changes
+
+-   Updated dependencies [[`e6f76c42d840389f791d2767de46b063bb392180`](https://github.com/learningeconomy/LearnCard/commit/e6f76c42d840389f791d2767de46b063bb392180)]:
+    -   @learncard/types@5.8.0
+    -   @learncard/core@9.3.31
+
 ## 1.0.40
 
 ### Patch Changes